\name{h.all.v2024.1.Hs.symbols.gmt}
\alias{hallmark gene sets}

\title{hallmark gene sets}

\description{
This is the hallmark gene sets collection from the Molecular Signatures Databas. 
}

\references{Liberzon et al. "The Molecular Signatures Database (MSigDB) hallmark gene set collection" \emph{Bioinformatics, Volume 27, Issue 12, June 2011, Pages 1739-1740} \doi{https://doi.org/10.1093/bioinformatics/btr260}}

\author{Stefano M. Pagnotta}
