\name{pre_ranked_list}
\alias{pre_ranked_list}

\title{FGFR3-TACC3 fusion positive gene profile}

\description{
This gene-profile comes from the paper in reference. It compares 9 FGFR3-TACC3 fusion positive samples versus 535 other samples in the GBM study from TCGA (Agilent platform). 
}

\references{Frattini et al. "A metabolic function of FGFR3-TACC3 gene fusions in cancer" \emph{Nature volume 553, 2018} \doi{10.1038/nature25171}}

\author{Stefano M. Pagnotta}

